working on memory leak in .vbs
