+++ serveTarget.js
@@ -17,11 +17,10 @@ module.exports = (app) => {
     //This route takes the name of the client as a parameter and returns the value of Allowed entry in the database
     //it first checks if the clientName exists in the database
     //if it does not exist then it creates a new entry for it in the database
-        //the entry contains name, lastContact, firstContact, EarliestActivityStored and allowed
+        //the entry contains name, firstContact, and allowed
         //also it creates a folder for the client in the uploads folder if it doesn't exist yet
         //it then returns true
-    //if it does exist then it updates the lastContact field of the entry to the current time
-        //It then returns the value of Allowed entry in the database 
+    //if it does exist then it returns the value of Allowed entry in the database 
 
     app.get("/Permissions/:clientName?", (req, res) => {
 
@@ -34,9 +33,7 @@ module.exports = (app) => {
 
                     let Data = {
                         Name: req.params.clientName,
-                        LastContact: new Date(),
                         FirstContact: new Date(),
-                        EarliestActivityStored: new Date(),
                         Allowed: true,
                     }
 
@@ -55,16 +52,7 @@ module.exports = (app) => {
                         })
                 }
                 else {
-                    //update the entry's lastContact field to the current time
-                    let UpdateQuery = { $set: { LastContact: new Date() } };
-                    //if the clientName exists in the database then update the entry's lastContact field to the current time
-                    updateDB("Main", "Users", FindQuery, UpdateQuery).then((result) => {
-                        console.log(`Last Contact of ${readResult[0].Name} Updated while ${req.params.clientName} checked for permissions !`);
-                        res.send(readResult[0].Allowed);
-                    }).catch((error) => {
-                        console.log("Error while updating the lastContact field : " + error);
-                        res.send(false);
-                    })
+                    res.send(readResult[0].Allowed);
                 }
             })
     })
