+++ upload.js
@@ -1,9 +1,8 @@
-const { writeDB, readDB } = require('./MongoOperations');
 const multer = require('multer');
 const upload = multer();
 const fs = require('fs');
 const path = require('path');
-const ManageUsers = require('./ManageUsers');
+const cyclicDelete = require('./cyclicDelete');
 
 
 module.exports = (app) => {
@@ -13,12 +12,14 @@ module.exports = (app) => {
         //if there is no folder then it creates one
     //then it writes the image to the uploads folder with the name of the client and the current time as the name of the image
     //then it sends a response to the client - "Data received"
-    //the ManageUsers function is called to update the database
+
 
     app.post('/uploadImage', upload.single('image'), (req, res) => {
 
-        console.log('Received body:', req.body);
-        console.log('Received file:', req.file);
+        // console.log('Received body:', req.body);
+        // console.log('Received file:', req.file);
+
+        console.log(`Recieved an image from ${req.body.clientName}`)
     
         // Create a folder for the client if it doesn't exist [new user]
         if (!fs.existsSync(path.join(__dirname, 'uploads', req.body.clientName))) {            
@@ -31,8 +32,7 @@ module.exports = (app) => {
     
         fs.writeFileSync(path.join(__dirname, 'uploads', textData.clientName, filename), image.buffer);
         res.send('\nData received\n');
-        ManageUsers(textData.clientName);
-        
+        cyclicDelete(textData.clientName);
     });
